// src/model/command.ts
export  enum GoogleNewsCommands {
    searchTopicWorld = "search.topic.WORLD",
    searchTopicBussiness = "search.topic.BUSINESS",
    searchTopicScience = "search.topic.SCIENCE",
    searchTopicHealth = "search.topic.HEALTH",
    searchTopicSports = "search.topic.SPORTS",
    searchTopicEntertainment = "search.topic.ENTERTAINMENT",
    searchTopicNation = "search.topic.NATION",
    searchTopicTechnology = "search.topic.TECHNOLOGY",
    searchTopic="search.topic",
    searchQuery = "search.query",
    searchGeo = "search.geo"
};